serverless/requirements.txt
functions-framework==3.*
google-cloud-storage==2.18.2
langchain-google-genai==2.0.1
langchain-core==0.3.12
langchain==0.3.4